further reduced lcp
