+++ package.json
@@ -23,6 +23,7 @@
   "devDependencies": {
     "@types/draftjs-to-html": "^0.8.1",
     "@types/html-to-draftjs": "^1.4.0",
+    "@types/loadable__component": "^5.13.9",
     "@types/lodash": "^4.14.180",
     "@types/react": "^17.0.38",
     "@types/react-avatar-editor": "^12.0.0",
@@ -69,6 +70,7 @@
     "@fortawesome/free-regular-svg-icons": "^6.1.2",
     "@fortawesome/free-solid-svg-icons": "^6.2.1",
     "@fortawesome/react-fontawesome": "^0.2.0",
+    "@loadable/component": "^5.16.4",
     "@testing-library/jest-dom": "^5.16.2",
     "@testing-library/react": "^12.1.3",
     "@testing-library/user-event": "^13.5.0",
