+++ src/nuance_assets/index.html
@@ -14,6 +14,26 @@
       rel="apple-touch-icon"
       href="https://7vltd-byaaa-aaaaf-qagqq-cai.raw.ic0.app/storage?contentId=a43wv-xyzwl-nzirb-3bqy3-qh5xr-5mvsy-j2hl4-ua566-fh6lj-xisav-zqe-image-5236"
     />
+    <link 
+      rel="preconnect"
+      href="https://fonts.googleapis.com"
+      crossorigin="anonymous"
+    />
+    <link 
+      rel="stylesheet"
+      href="https://fonts.googleapis.com"
+      crossorigin="anonymous"
+    />
+    <link 
+      rel="preconnect"
+      href="https://fonts.gstatic.com"
+      crossorigin="anonymous"
+    />
+    <link 
+      rel="stylesheet"
+      href="https://fonts.gstatic.com"
+      crossorigin="anonymous"
+    />
     <meta
       name="google-site-verification"
       content="qMZShFZjZvGGBZFC09yptL-Zn5yncOdtZp1DPstb0BI"
